UAG - Cleanup - Advanced columns.
