+++ src/blocks/column/style.scss
@@ -3,30 +3,35 @@
  *
  * CSS for both Frontend+Backend.
  */
+ .uagb-column__wrap,
+ .uagb-column__inner-wrap {
+	margin-left: auto;
+	margin-right: auto;
+	position: relative;
+	width: 100%;
+	z-index: 2;
+}
 
 .uagb-column__wrap {
 
 	position: relative;
 	overflow:hidden;
 
-	.uagb-column__inner-wrap {
-		margin-left: auto;
-		margin-right: auto;
-		position: relative;
+	> * {
 		z-index: 2;
-		width: 100%;
+		padding: 5px;
 	}
-
-	&.uagb-column__align-left	.uagb-column__inner-wrap {
+	
+	&.uagb-column__align-left {
 		margin-left: 0;
 		margin-right: auto;
 	}
 
-	&.uagb-column__align-right .uagb-column__inner-wrap {
+	&.uagb-column__align-right {
 		margin-left: auto;
 		margin-right: 0;
 	}
-
+	.uagb-column__video-wrap,
 	.uagb-column__overlay {
 		height: 100%;
 		width: 100%;
@@ -36,11 +41,6 @@
 	}
 
 	.uagb-column__video-wrap {
-		height: 100%;
-		width: 100%;
-		top: 0;
-		left: 0;
-		position: absolute;
 		overflow: hidden;
 		z-index: 0;
 		-webkit-transition: opacity 1s;
@@ -78,24 +78,24 @@
 }
 
 @media (max-width: 976px) {
-	.uagb-column__align-tablet-left .uagb-column__inner-wrap {
+	.uagb-column__align-tablet-left {
 		margin-left: 0;
 		margin-right: auto;
 	}
 
-	.uagb-column__align-tablet-right .uagb-column__inner-wrap {
+	.uagb-column__align-tablet-right {
 		margin-left: auto;
 		margin-right: 0;
 	}
 }
 
 @media (max-width: 767px) {
-	.uagb-column__align-mobile-left .uagb-column__inner-wrap {
+	.uagb-column__align-mobile-left {
 		margin-left: 0;
 		margin-right: auto;
 	}
 
-	.uagb-column__align-mobile-right .uagb-column__inner-wrap {
+	.uagb-column__align-mobile-right {
 		margin-left: auto;
 		margin-right: 0;
 	}
