+++ src/blocks/columns/attributes.js
@@ -2,7 +2,7 @@
  * BLOCK: Columns - Attributes
  */
 
-const attributes = {
+ const attributes = {
 	block_id: {
 		type: "string",
 	},
@@ -257,27 +257,27 @@ const attributes = {
 	},
 	mobileMarginType: {
 		type: "string",
-		default: 'px'
+		default: "px"
 	},
 	tabletMarginType: {
 		type: "string",
-		default: 'px'
+		default: "px"
 	},
 	desktopMarginType: {
 		type: "string",
-		default: 'px'
+		default: "px"
 	},
 	mobilePaddingType: {
 		type: "string",
-		default: 'px'
+		default: "px"
 	},
 	tabletPaddingType: {
 		type: "string",
-		default: 'px'
+		default: "px"
 	},
 	desktopPaddingType: {
 		type: "string",
-		default: 'px'
+		default: "px"
 	},
 	boxShadowColor: {
 		type: "string"
