+++ src/blocks/columns/icons.js
@@ -2,71 +2,71 @@
  * WordPress dependencies
  */
 
-const { Path, SVG, G } = wp.components
-/**
- * Custom icons
- */
-const rowIcons = {};
-
-rowIcons.colOne =
-	<SVG className="dashicon" height="26" viewBox="0 0 50 26" width="50" xmlns="http://www.w3.org/2000/svg"><Path d="m48.0833333 0h-46.16666663c-1.05416667 0-1.91666667.9-1.91666667 2v22c0 1.1.8625 2 1.91666667 2h46.16666663c1.0541667 0 1.9166667-.9 1.9166667-2v-22c0-1.1-.8625-2-1.9166667-2zm0 24h-46.16666663v-22h46.16666663z" /></SVG>;
-
-rowIcons.layout5050 =
-	<SVG className="dashicon" height="26" viewBox="0 0 50 26" width="50" xmlns="http://www.w3.org/2000/svg"><G fillRule="evenodd"><Path d="m48.0833333 0h-46.16666663c-1.05416667 0-1.91666667.9-1.91666667 2v22c0 1.1.8625 2 1.91666667 2h46.16666663c1.0541667 0 1.9166667-.9 1.9166667-2v-22c0-1.1-.8625-2-1.9166667-2zm0 24h-46.16666663v-22h46.16666663z" fillRule="nonzero" /><Path d="m24 2h2v22h-2z" /></G></SVG>;
-
-rowIcons.layout502525 =
-	<SVG className="dashicon" height="26" viewBox="0 0 50 26" width="50" xmlns="http://www.w3.org/2000/svg"><G fillRule="evenodd"><Path d="m48.0833333 0h-46.16666663c-1.05416667 0-1.91666667.9-1.91666667 2v22c0 1.1.8625 2 1.91666667 2h46.16666663c1.0541667 0 1.9166667-.9 1.9166667-2v-22c0-1.1-.8625-2-1.9166667-2zm0 24h-46.16666663v-22h46.16666663z" fillRule="nonzero" /><Path d="m36 2h2v22h-2z" /><Path d="m24 2h2v22h-2z" /></G></SVG>;
-
-rowIcons.layout25252525 =
-	<SVG className="dashicon" height="26" viewBox="0 0 50 26" width="50" xmlns="http://www.w3.org/2000/svg"><G fillRule="evenodd"><Path d="m48.0833333 0h-46.16666663c-1.05416667 0-1.91666667.9-1.91666667 2v22c0 1.1.8625 2 1.91666667 2h46.16666663c1.0541667 0 1.9166667-.9 1.9166667-2v-22c0-1.1-.8625-2-1.9166667-2zm0 24h-46.16666663v-22h46.16666663z" fillRule="nonzero" /><Path d="m12 2h2v22h-2z" /><Path d="m24 2h2v22h-2z" /><Path d="m36 2h2v22h-2z" /></G></SVG>;
-
-rowIcons.layout3366 = 
-<SVG width="48" height="48" viewBox="0 0 48 48" xmlns="http://www.w3.org/2000/svg">
-<Path
-	fillRule="evenodd"
-	clipRule="evenodd"
-	d="M39 12C40.1046 12 41 12.8954 41 14V34C41 35.1046 40.1046 36 39 36H9C7.89543 36 7 35.1046 7 34V14C7 12.8954 7.89543 12 9 12H39ZM39 34V14H20V34H39ZM18 34H9V14H18V34Z"
-/>
-</SVG>
-
-rowIcons.layout6633 = 
-<SVG
-width="48"
-height="48"
-viewBox="0 0 48 48"
-xmlns="http://www.w3.org/2000/svg"
->
-<Path
-	fillRule="evenodd"
-	clipRule="evenodd"
-	d="M39 12C40.1046 12 41 12.8954 41 14V34C41 35.1046 40.1046 36 39 36H9C7.89543 36 7 35.1046 7 34V14C7 12.8954 7.89543 12 9 12H39ZM39 34V14H30V34H39ZM28 34H9V14H28V34Z"
-/>
-</SVG>
-
-rowIcons.layout333333 = 
-<SVG
-width="48"
-height="48"
-viewBox="0 0 48 48"
-xmlns="http://www.w3.org/2000/svg"
->
-<Path
-	fillRule="evenodd"
-	d="M41 14a2 2 0 0 0-2-2H9a2 2 0 0 0-2 2v20a2 2 0 0 0 2 2h30a2 2 0 0 0 2-2V14zM28.5 34h-9V14h9v20zm2 0V14H39v20h-8.5zm-13 0H9V14h8.5v20z"
-/>
-</SVG>
-
-rowIcons.layout255025 = 
-<SVG
-width="48"
-height="48"
-viewBox="0 0 48 48"
-xmlns="http://www.w3.org/2000/svg"
->
-<Path
-	fillRule="evenodd"
-	d="M41 14a2 2 0 0 0-2-2H9a2 2 0 0 0-2 2v20a2 2 0 0 0 2 2h30a2 2 0 0 0 2-2V14zM31 34H17V14h14v20zm2 0V14h6v20h-6zm-18 0H9V14h6v20z"
-/>
-</SVG>
-
-export default rowIcons;
+ const { Path, SVG, G } = wp.components
+ /**
+  * Custom icons
+  */
+ const rowIcons = {}
+ 
+ rowIcons.colOne =
+	 <SVG className="dashicon" height="26" viewBox="0 0 50 26" width="50" xmlns="http://www.w3.org/2000/svg"><Path d="m48.0833333 0h-46.16666663c-1.05416667 0-1.91666667.9-1.91666667 2v22c0 1.1.8625 2 1.91666667 2h46.16666663c1.0541667 0 1.9166667-.9 1.9166667-2v-22c0-1.1-.8625-2-1.9166667-2zm0 24h-46.16666663v-22h46.16666663z" /></SVG>
+ 
+ rowIcons.layout5050 =
+	 <SVG className="dashicon" height="26" viewBox="0 0 50 26" width="50" xmlns="http://www.w3.org/2000/svg"><G fillRule="evenodd"><Path d="m48.0833333 0h-46.16666663c-1.05416667 0-1.91666667.9-1.91666667 2v22c0 1.1.8625 2 1.91666667 2h46.16666663c1.0541667 0 1.9166667-.9 1.9166667-2v-22c0-1.1-.8625-2-1.9166667-2zm0 24h-46.16666663v-22h46.16666663z" fillRule="nonzero" /><Path d="m24 2h2v22h-2z" /></G></SVG>
+ 
+ rowIcons.layout502525 =
+	 <SVG className="dashicon" height="26" viewBox="0 0 50 26" width="50" xmlns="http://www.w3.org/2000/svg"><G fillRule="evenodd"><Path d="m48.0833333 0h-46.16666663c-1.05416667 0-1.91666667.9-1.91666667 2v22c0 1.1.8625 2 1.91666667 2h46.16666663c1.0541667 0 1.9166667-.9 1.9166667-2v-22c0-1.1-.8625-2-1.9166667-2zm0 24h-46.16666663v-22h46.16666663z" fillRule="nonzero" /><Path d="m36 2h2v22h-2z" /><Path d="m24 2h2v22h-2z" /></G></SVG>
+ 
+ rowIcons.layout25252525 =
+	 <SVG className="dashicon" height="26" viewBox="0 0 50 26" width="50" xmlns="http://www.w3.org/2000/svg"><G fillRule="evenodd"><Path d="m48.0833333 0h-46.16666663c-1.05416667 0-1.91666667.9-1.91666667 2v22c0 1.1.8625 2 1.91666667 2h46.16666663c1.0541667 0 1.9166667-.9 1.9166667-2v-22c0-1.1-.8625-2-1.9166667-2zm0 24h-46.16666663v-22h46.16666663z" fillRule="nonzero" /><Path d="m12 2h2v22h-2z" /><Path d="m24 2h2v22h-2z" /><Path d="m36 2h2v22h-2z" /></G></SVG>
+ 
+ rowIcons.layout3366 = 
+ <SVG width="48" height="48" viewBox="0 0 48 48" xmlns="http://www.w3.org/2000/svg">
+	 <Path
+		 fillRule="evenodd"
+		 clipRule="evenodd"
+		 d="M39 12C40.1046 12 41 12.8954 41 14V34C41 35.1046 40.1046 36 39 36H9C7.89543 36 7 35.1046 7 34V14C7 12.8954 7.89543 12 9 12H39ZM39 34V14H20V34H39ZM18 34H9V14H18V34Z"
+	 />
+ </SVG>
+ 
+ rowIcons.layout6633 = 
+ <SVG
+	 width="48"
+	 height="48"
+	 viewBox="0 0 48 48"
+	 xmlns="http://www.w3.org/2000/svg"
+ >
+	 <Path
+		 fillRule="evenodd"
+		 clipRule="evenodd"
+		 d="M39 12C40.1046 12 41 12.8954 41 14V34C41 35.1046 40.1046 36 39 36H9C7.89543 36 7 35.1046 7 34V14C7 12.8954 7.89543 12 9 12H39ZM39 34V14H30V34H39ZM28 34H9V14H28V34Z"
+	 />
+ </SVG>
+ 
+ rowIcons.layout333333 = 
+ <SVG
+	 width="48"
+	 height="48"
+	 viewBox="0 0 48 48"
+	 xmlns="http://www.w3.org/2000/svg"
+ >
+	 <Path
+		 fillRule="evenodd"
+		 d="M41 14a2 2 0 0 0-2-2H9a2 2 0 0 0-2 2v20a2 2 0 0 0 2 2h30a2 2 0 0 0 2-2V14zM28.5 34h-9V14h9v20zm2 0V14H39v20h-8.5zm-13 0H9V14h8.5v20z"
+	 />
+ </SVG>
+ 
+ rowIcons.layout255025 = 
+ <SVG
+	 width="48"
+	 height="48"
+	 viewBox="0 0 48 48"
+	 xmlns="http://www.w3.org/2000/svg"
+ >
+	 <Path
+		 fillRule="evenodd"
+		 d="M41 14a2 2 0 0 0-2-2H9a2 2 0 0 0-2 2v20a2 2 0 0 0 2 2h30a2 2 0 0 0 2-2V14zM31 34H17V14h14v20zm2 0V14h6v20h-6zm-18 0H9V14h6v20z"
+	 />
+ </SVG>
+ 
+ export default rowIcons
