+++ src/blocks/columns/inline-styles.js
@@ -1,60 +1,60 @@
 /**
  * Returns Dynamic Generated CSS
  */
-import generateCSSUnit from "../../../dist/blocks/uagb-controls/generateCSSUnit"
+ import generateCSSUnit from "../../../dist/blocks/uagb-controls/generateCSSUnit"
 
-function inlineStyles( props, isEditor ) {
-
-	const {
-		leftPadding,
-		rightPadding,
-		topPadding,
-		bottomPadding,
-		topMargin,
-		bottomMargin,
-		backgroundPosition,
-		backgroundSize,
-		backgroundAttachment,
-		backgroundImage,
-		backgroundRepeat,
-		backgroundType,
-		borderStyle,
-		borderWidth,
-		borderRadius,
-		borderColor,
-		desktopMarginType,
-		desktopPaddingType,
-	} = props.attributes
-
-	var style = {
-		"padding-top": generateCSSUnit( topPadding , desktopPaddingType),
-		"padding-bottom": generateCSSUnit( bottomPadding , desktopPaddingType),
-		"padding-left": generateCSSUnit( leftPadding , desktopPaddingType),
-		"padding-right": generateCSSUnit( rightPadding , desktopPaddingType),
-		"margin-top": generateCSSUnit( topMargin , desktopMarginType),
-		"margin-bottom": generateCSSUnit( bottomMargin , desktopMarginType),
-		"border-radius": generateCSSUnit( borderRadius , desktopMarginType),
-	}
-
-	if ( borderStyle != "none" ) {
-		style["border-style"] = borderStyle
-		style["border-width"] = generateCSSUnit( borderWidth, "px" )
-		style["border-color"] =  borderColor
-	}
-
-	var position = backgroundPosition.replace( "-", " " )
-
-	if ( "image" === backgroundType ) {
-
-		style["background-image"] = ( backgroundImage ) ? `url(${ backgroundImage.url })` : null
-		style["background-position"] = position
-		style["background-attachment"] = backgroundAttachment
-		style["background-repeat"] = backgroundRepeat
-		style["background-size"] = backgroundSize
-
-	}
-
-	return style
-}
-
-export default inlineStyles
+ function inlineStyles( props, isEditor ) {
+ 
+	 const {
+		 leftPadding,
+		 rightPadding,
+		 topPadding,
+		 bottomPadding,
+		 topMargin,
+		 bottomMargin,
+		 backgroundPosition,
+		 backgroundSize,
+		 backgroundAttachment,
+		 backgroundImage,
+		 backgroundRepeat,
+		 backgroundType,
+		 borderStyle,
+		 borderWidth,
+		 borderRadius,
+		 borderColor,
+		 desktopMarginType,
+		 desktopPaddingType,
+	 } = props.attributes
+ 
+	 var style = {
+		 "padding-top": generateCSSUnit( topPadding , desktopPaddingType ),
+		 "padding-bottom": generateCSSUnit( bottomPadding , desktopPaddingType ),
+		 "padding-left": generateCSSUnit( leftPadding , desktopPaddingType ),
+		 "padding-right": generateCSSUnit( rightPadding , desktopPaddingType ),
+		 "margin-top": generateCSSUnit( topMargin , desktopMarginType ),
+		 "margin-bottom": generateCSSUnit( bottomMargin , desktopMarginType ),
+		 "border-radius": generateCSSUnit( borderRadius , desktopMarginType ),
+	 }
+ 
+	 if ( borderStyle != "none" ) {
+		 style["border-style"] = borderStyle
+		 style["border-width"] = generateCSSUnit( borderWidth, "px" )
+		 style["border-color"] =  borderColor
+	 }
+ 
+	 var position = backgroundPosition.replace( "-", " " )
+ 
+	 if ( "image" === backgroundType ) {
+ 
+		 style["background-image"] = ( backgroundImage ) ? `url(${ backgroundImage.url })` : null
+		 style["background-position"] = position
+		 style["background-attachment"] = backgroundAttachment
+		 style["background-repeat"] = backgroundRepeat
+		 style["background-size"] = backgroundSize
+ 
+	 }
+ 
+	 return style
+ }
+ 
+ export default inlineStyles
